+++ examples/plot_quickstart_guide.py
@@ -639,5 +639,58 @@ plot_plan(P_fgw, "Fused GW plan", axis=False)
 pl.show()
 
 # sphinx_gallery_end_ignore
+
+# %%
+#
+# Solving barycenter problems
+# -------------------------------
+# Solve Optimal transport barycenter problem with free support between several input distributions.
+# ~~~~~~~~~~~~~~~~~~~~
+#
+# The :func:`ot.solve_bary_sample` function can be used to solve the Optimal Transport barycenter problem
+# between multiple sets of samples while optimizing the support of the barycenter and letting fixed their probability weights.
+# The function takes as its first argument the list of samples in each input distribution,
+# and as second argument the number of samples to learn in the barycenter. By default, the probability weights in each distribution and the barycentric weights are uniform but they can be customized by the user.
+#
+# The function returns an :class:`ot.utils.OTBaryResult` object that contains in part the barycenter samples and the OT plans between the barycenter and each input distribution.
+#
+# In the following, we illustrate the use of this function with the same 2D data as above considered as input distributions and compute their barycenter while using exact OT.
+# Notice that most of the arguments of the :func:`ot.solve_bary_sample` function are similar to those of the :func:`ot.solve_sample` function and that the same regularization and unbalanced parameters can be used to solve regularized and unbalanced barycenter problems.
+
+# Solve the OT barycenter problem (exact OT without any regularization)
+sol = ot.solve_bary_sample([x1, x2], n=35)
+
+# get the barycenter support
+X = sol.X
+
+# get the OT plans between the barycenter and each input distribution
+list_P = [sol.list_res[i].plan for i in range(2)]
+
+# get the barycenterOT loss
+loss = sol.value
+
+print(f"Barycenter OT loss = {loss:1.3f}")
+
+# sphinx_gallery_start_ignore
+pl.figure(1, (8, 8))
+plot2D_samples_mat(x1, X, list_P[0])
+plot2D_samples_mat(x2, X, list_P[1])
+
+pl.plot(x1[:, 0], x1[:, 1], "ob", label="Source distribution 1", **style)
+pl.plot(x2[:, 0], x2[:, 1], "or", label="Source distribution 2", **style)
+pl.plot(X[:, 0], X[:, 1], "og", label="Barycenter distribution", **style)
+
+pl.title(
+    "Barycenter samples and OT plans \n total loss= %s = 0.5 * %s + 0.5 * %s"
+    % (
+        np.round(loss, 3),
+        np.round(sol.list_res[0].value, 3),
+        np.round(sol.list_res[1].value, 3),
+    )
+)
+pl.legend(loc="best")
+pl.show()
+
+# sphinx_gallery_end_ignore
 # %%
 #